Unify climb info page with play drawer UI
