+++ packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/view/[climb_uuid]/page.tsx
@@ -132,12 +132,7 @@ export default async function DynamicResultsPage(props: { params: Promise<BoardR
       <ClimbDetailPageServer
         climb={climbWithProcessedData}
         boardDetails={boardDetails}
-        betaLinks={detailData.betaLinks}
-        climbUuid={parsedParams.climb_uuid}
-        boardType={parsedParams.board_name}
         angle={parsedParams.angle}
-        currentClimbDifficulty={currentClimb.difficulty ?? undefined}
-        boardName={parsedParams.board_name}
       />
     );
   } catch (error) {
