+++ packages/web/app/components/climb-detail/climb-detail-page.server.tsx
@@ -1,62 +1,17 @@
 import React from 'react';
-import ClimbViewActions from '@/app/components/climb-view/climb-view-actions';
-import ClimbDetailInfoShellClient from '@/app/components/climb-detail/climb-detail-info-shell.client';
-import { constructClimbInfoUrl } from '@/app/lib/url-utils';
-import type { BetaLink } from '@/app/lib/api-wrappers/sync-api-types';
 import type { BoardDetails, Climb } from '@/app/lib/types';
-import styles from '@/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/view/[climb_uuid]/climb-view.module.css';
+import PlayStyleClimbDetailClient from '@/app/components/climb-detail/play-style-climb-detail.client';
 
 interface ClimbDetailPageServerProps {
   climb: Climb;
   boardDetails: BoardDetails;
-  betaLinks: BetaLink[];
-  climbUuid: string;
-  boardType: string;
   angle: number;
-  currentClimbDifficulty?: string;
-  boardName?: string;
 }
 
 export default function ClimbDetailPageServer({
   climb,
   boardDetails,
-  betaLinks,
-  climbUuid,
-  boardType,
   angle,
-  currentClimbDifficulty,
-  boardName,
 }: ClimbDetailPageServerProps) {
-  const auroraAppUrl = constructClimbInfoUrl(
-    boardDetails,
-    climb.uuid,
-  ) ?? undefined;
-
-  return (
-    <div className={styles.pageContainer}>
-      <div className={styles.actionsSection}>
-        <ClimbViewActions
-          climb={climb}
-          boardDetails={boardDetails}
-          auroraAppUrl={auroraAppUrl}
-          angle={angle}
-        />
-      </div>
-
-      <div className={styles.contentWrapper}>
-        <div className={styles.climbSection}>
-          <ClimbDetailInfoShellClient
-            climb={climb}
-            boardDetails={boardDetails}
-            betaLinks={betaLinks}
-            climbUuid={climbUuid}
-            boardType={boardType}
-            angle={angle}
-            currentClimbDifficulty={currentClimbDifficulty}
-            boardName={boardName}
-          />
-        </div>
-      </div>
-    </div>
-  );
+  return <PlayStyleClimbDetailClient climb={climb} boardDetails={boardDetails} angle={angle} />;
 }
